Performance increase (negligible). Removed random type conversions, shaved off 3-5 milliseconds in release mode.

+++ src/generate_password.rs
@@ -1,6 +1,6 @@
 use rand::Rng;
 
-static ITERATION_RANGE: u32 = 4_000_000;
+static ITERATION_RANGE: usize = 4_000_000;
 
 const STANDARD: &[char] = &[
     'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r',
@@ -36,7 +36,7 @@ pub fn generate_passwords(length: u8, num_pws: u8, ext_special: bool) -> u8 {
         let mut password: Vec<char> = Vec::<char>::new();
         for _j in 0..length {
             let rand_selection = rand::thread_rng().gen_range(0..ITERATION_RANGE);
-            let char_num = random_ints[rand_selection as usize];
+            let char_num = random_ints[rand_selection];
             let rand_char = get_char(char_num);
             password.push(rand_char);
         }
